+++ apps/api/src/module/user/controller/user.resolver.ts
@@ -1,6 +1,6 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { UserObject } from './dto/object/user-object';
-import { RoomObject } from '@/module/room/controller/dto/object/room-object';
+import { UserObject } from './dto/object/user.object';
+import { RoomObject } from '@/module/room/controller/dto/object/room.object';
 import { RoomDataLoader } from '@/module/room/dataloader/room.dataloader';
 import type { Room } from '@/module/room/domain/room.model';
 import { User } from '@/module/user/domain/user.model';
